Move some Rucio utils over to RucioUtils module
fix source docstring

Replace custom weight choice by the random library
