+++ test/python/WMCore_t/Services_t/Rucio_t/RucioUtils_t.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python
+"""
+Unittests for ...
+"""
+import unittest
+
+class MyTestCase(unittest.TestCase):
+    def test_something(self):
+        self.assertEqual(True, False)
+
+if __name__ == '__main__':
+    unittest.main()
